QoL stuff
Did some polishing on farms. Moved back to modals on farms rather than accordions. Having issues pulling farm.stakedUSD. Able to put checks and filter it out when not connected, but if you connect, disconnect, and then connect again, it crashes. Trying to find a clever way to pop them into the modals or find a workaround for the FarmList.